requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap5==2024.2
cuser==0.0.4
discord.py==2.4.0
Django==4.2.13
django-abstract-utils==0.0.1
django-crispy-forms==2.3
django-cuser==2014.9.28
django-jazzmin==3.0.0
django-minimal-abstract-user==0.0.2
django-modeltranslation==0.19.5
django-phonenumber-field==8.0.0
django-tinymce==4.1.0
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
modeltranslation==0.25
multidict==6.0.5
packaging==24.1
paginator.py==0.9.4
phonenumbers==8.13.45
Pillow==10.0.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pypng==0.20220715.0
pyserial==3.5
python-gettext==5.0
python-i18n==0.3.9
qrcode==7.4.2
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
